Follow slv2 changes (size_t -> unsigned int)
git-svn-id: svn://svn.gna.org/svn/zynjacku/code@58 c18b01a0-d024-0410-a44b-803e0d4acf19

+++ synth.c
@@ -760,7 +760,7 @@ zynjacku_synth_construct(
   guint sample_rate;
   struct zynjacku_engine * engine_ptr;
   SLV2Strings slv2_strings;
-  size_t string_index;
+  unsigned int string_index;
   gboolean dynparams_supported;
 
   synth_ptr = ZYNJACKU_SYNTH_GET_PRIVATE(synth_obj_ptr);
